Change the internal process running logic to handle redirecting output/error async.
Rather than polling a processes output/error streams, use the async events and block until we have read everything.
This fixes cases where the running process leaves a stream open until the process exits, resulting in a deadlock.

A new start info property of `OutputRedirectingTimeout` has been added, which is -1 by default (meaning block indefinitely), otherwise
it is the number of milliseconds to wait for a redirecting output/error process to close the stream.